+++ examples/testing/emails.php
@@ -1,14 +1,9 @@
 <?php
 
-use Mailtrap\Config;
-use Mailtrap\EmailHeader\CategoryHeader;
-use Mailtrap\EmailHeader\CustomVariableHeader;
-use Mailtrap\EmailHeader\Template\TemplateUuidHeader;
-use Mailtrap\EmailHeader\Template\TemplateVariableHeader;
 use Mailtrap\Helper\ResponseHelper;
 use Mailtrap\MailtrapClient;
+use Mailtrap\Mime\MailtrapEmail;
 use Symfony\Component\Mime\Address;
-use Symfony\Component\Mime\Email;
 use Symfony\Component\Mime\Header\UnstructuredHeader;
 
 require __DIR__ . '/../vendor/autoload.php';
@@ -25,12 +20,14 @@ require __DIR__ . '/../vendor/autoload.php';
  * POST https://sandbox.api.mailtrap.io/api/send/{inbox_id}
  */
 try {
-    // your API token from here https://mailtrap.io/api-tokens
-    $apiKey = getenv('MAILTRAP_API_KEY');
-    $mailtrap = new MailtrapClient(new Config($apiKey));
+    $mailtrap = MailtrapClient::initSendingEmails(
+        apiKey: getenv('MAILTRAP_API_KEY'), #your API token from here https://mailtrap.io/api-tokens
+        isSandbox: true, # Sandbox sending (@see https://help.mailtrap.io/article/109-getting-started-with-mailtrap-email-testing)
+        inboxId: getenv('MAILTRAP_INBOX_ID') # required param for sandbox sending
+    );
 
-    $email = (new Email())
-        ->from(new Address('mailtrap@example.com', 'Mailtrap Test'))
+    $email = (new MailtrapEmail())
+        ->from(new Address('mailtrap@example.com', 'Mailtrap Test')) // <--- you should use the domain which is linked to template UUID (otherwise you will get 401)
         ->replyTo(new Address('reply@example.com'))
         ->to(new Address('email@example.com', 'Jon'))
         ->cc('mailtrapqa@example.com')
@@ -50,27 +47,20 @@ try {
         )
         ->embed(fopen('https://mailtrap.io/wp-content/uploads/2021/04/mailtrap-new-logo.svg', 'r'), 'logo', 'image/svg+xml')
         ->attachFromPath('README.md')
+        ->customVariables([
+            'user_id' => '45982',
+            'batch_id' => 'PSJ-12'
+        ])
+        ->category('Integration Test')
     ;
 
-    // Headers
+    // Custom email headers (optional)
     $email->getHeaders()
-        ->addTextHeader('X-Message-Source', '1alf.com')
+        ->addTextHeader('X-Message-Source', 'test.com')
         ->add(new UnstructuredHeader('X-Mailer', 'Mailtrap PHP Client'))
     ;
 
-    // Custom Variables
-    $email->getHeaders()
-        ->add(new CustomVariableHeader('user_id', '45982'))
-        ->add(new CustomVariableHeader('batch_id', 'PSJ-12'))
-    ;
-
-    // Category (should be only one)
-    $email->getHeaders()
-        ->add(new CategoryHeader('Integration Test'))
-    ;
-
-    // Required param -> inbox_id
-    $response = $mailtrap->sandbox()->emails()->send($email, 1000001); // <--- you should use your inbox_id here (otherwise you will get 401)
+    $response = $mailtrap->send($email);
 
     // print all possible information from the response
     var_dump($response->getHeaders()); //headers (array)
@@ -90,28 +80,30 @@ try {
  * Optional template variables that will be used to generate actual subject, text and html from email template
  */
 try {
-    // your API token from here https://mailtrap.io/api-tokens
-    $apiKey = getenv('MAILTRAP_API_KEY');
-    $mailtrap = new MailtrapClient(new Config($apiKey));
+    $mailtrap = MailtrapClient::initSendingEmails(
+        apiKey: getenv('MAILTRAP_API_KEY'), #your API token from here https://mailtrap.io/api-tokens
+        isSandbox: true, # Sandbox sending (@see https://help.mailtrap.io/article/109-getting-started-with-mailtrap-email-testing)
+        inboxId: getenv('MAILTRAP_INBOX_ID') # required param for sandbox sending
+    );
 
-    $email = (new Email())
+    $email = (new MailtrapEmail())
         ->from(new Address('example@YOUR-DOMAIN-HERE.com', 'Mailtrap Test')) // <--- you should use the domain which is linked to template UUID (otherwise you will get 401)
         ->replyTo(new Address('reply@YOUR-DOMAIN-HERE.com'))
         ->to(new Address('example@gmail.com', 'Jon'))
+        ->templateUuid('bfa432fd-0000-0000-0000-8493da283a69')
+        ->templateVariables([
+            'user_name' => 'Jon Bush',
+            'next_step_link' => 'https://mailtrap.io/',
+            'get_started_link' => 'https://mailtrap.io/',
+            'onboarding_video_link' => 'some_video_link'
+        ])
     ;
 
-    // Template UUID and Variables
-    $email->getHeaders()
-        ->add(new TemplateUuidHeader('bfa432fd-0000-0000-0000-8493da283a69'))
-        ->add(new TemplateVariableHeader('user_name', 'Jon Bush'))
-        ->add(new TemplateVariableHeader('next_step_link', 'https://mailtrap.io/'))
-        ->add(new TemplateVariableHeader('get_started_link', 'https://mailtrap.io/'))
-        ->add(new TemplateVariableHeader('onboarding_video_link', 'some_video_link'))
-    ;
+    $response = $mailtrap->send($email);
 
-    // Required param -> inbox_id
-    $response = $mailtrap->sandbox()->emails()->send($email, 1000001); // <--- you should use your inbox_id here (otherwise you will get 401)
-
+    // print all possible information from the response
+    var_dump($response->getHeaders()); //headers (array)
+    var_dump($response->getStatusCode()); //status code (int)
     var_dump(ResponseHelper::toArray($response)); // body (array)
 } catch (Exception $e) {
     echo 'Caught exception: ',  $e->getMessage(), "\n";