+++ src/Api/General/Permission.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Mailtrap\Api\General;
 
 use Mailtrap\Api\AbstractApi;
+use Mailtrap\ConfigInterface;
 use Mailtrap\DTO\Request\Permission\Permissions;
 use Mailtrap\Exception\RuntimeException;
 use Psr\Http\Message\ResponseInterface;
@@ -14,17 +15,20 @@ use Psr\Http\Message\ResponseInterface;
  */
 class Permission extends AbstractApi implements GeneralInterface
 {
+    public function __construct(ConfigInterface $config, private int $accountId)
+    {
+        parent::__construct($config);
+    }
+
     /**
      * Get all resources in your account (Inboxes, Projects, Domains, Billing and Account itself) to which the token has admin access.
      *
-     * @param int $accountId
-     *
      * @return ResponseInterface
      */
-    public function getResources(int $accountId): ResponseInterface
+    public function getResources(): ResponseInterface
     {
         return $this->handleResponse($this->httpGet(
-            sprintf('%s/api/accounts/%s/permissions/resources', $this->getHost(), $accountId)
+            sprintf('%s/api/accounts/%s/permissions/resources', $this->getHost(), $this->getAccountId())
         ));
     }
 
@@ -33,19 +37,23 @@ class Permission extends AbstractApi implements GeneralInterface
      * If you send a combination of resource_type and resource_id that already exists, the permission is updated.
      * If the combination doesn’t exist, the permission is created.
      *
-     * @param int         $accountId
      * @param int         $accountAccessId
      * @param Permissions $permissions
      *
      * @return ResponseInterface
      */
-    public function update(int $accountId, int $accountAccessId, Permissions $permissions): ResponseInterface
+    public function update(int $accountAccessId, Permissions $permissions): ResponseInterface
     {
         return $this->handleResponse($this->httpPut(
-            sprintf('%s/api/accounts/%s/account_accesses/%s/permissions/bulk', $this->getHost(), $accountId, $accountAccessId),
+            sprintf('%s/api/accounts/%s/account_accesses/%s/permissions/bulk', $this->getHost(), $this->getAccountId(), $accountAccessId),
             [],
             ['permissions' => $this->getPayload($permissions)]
         ));
+    }
+
+    public function getAccountId(): int
+    {
+        return $this->accountId;
     }
 
     private function getPayload(Permissions $permissions): array