+++ src/HttpClient/HttpClientBuilderInterface.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Mailtrap\HttpClient;
 
 use Psr\Http\Client\ClientInterface;