+++ src/MailtrapSendingClient.php
@@ -4,14 +4,12 @@ declare(strict_types=1);
 
 namespace Mailtrap;
 
-use Mailtrap\Api;
-
 /**
- * @method  Api\Sending\Emails   emails
+ * @method  Api\Sending\Emails   emails()
  *
  * Class MailtrapSendingClient
  */
-final class MailtrapSendingClient extends AbstractMailtrapClient
+final class MailtrapSendingClient extends AbstractMailtrapClient implements EmailsSendMailtrapClientInterface
 {
     public const API_MAPPING = [
         'emails' => Api\Sending\Emails::class,