+++ tests/MailtrapGeneralClientTest.php
@@ -26,8 +26,11 @@ class MailtrapGeneralClientTest extends MailtrapClientTestCase
 
     public function mapInstancesProvider(): iterable
     {
-        foreach (MailtrapGeneralClient::API_MAPPING as $item) {
-            yield [new $item($this->getConfigMock())];
+        foreach (MailtrapGeneralClient::API_MAPPING as $key => $item) {
+            yield match ($key) {
+                'permissions', 'users' => [new $item($this->getConfigMock(), self::FAKE_ACCOUNT_ID)],
+                default => [new $item($this->getConfigMock())],
+            };
         }
     }
 }